+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Change elm-verify-examples.json to contain elm root and not source root. In addition it will respect source-directories defined in elm.json [#109](https://github.com/stoeffel/elm-verify-examples/pull/109)
 
+    To migrate, replace the "root" key in your config file for a "elm-root" pointing to the folder containing your elm.json file. Typically this will look like this:
+    
+     ```diff
+     - "root": "../src",
+     + "elm-root": "..",
+     ```
+
 ## 5.3.0
 
 - Do not run tests by default, only generate [#65](https://github.com/stoeffel/elm-verify-examples/issues/65)